Module fedimint_ln::contracts

Modules

account

incoming

outgoing

Macros

try ğŸ”’

try ğŸ”’

try ğŸ”’

try ğŸ”’

try ğŸ”’

try ğŸ”’

try ğŸ”’

try ğŸ”’

try ğŸ”’

try ğŸ”’

try ğŸ”’

try ğŸ”’

try ğŸ”’

try ğŸ”’

try ğŸ”’

try ğŸ”’

try ğŸ”’

try ğŸ”’

Structs

AccountContractOutcome

ContractId

The hash of a LN incoming contract

EncryptedPreimage

Threshold-encrypted Preimage

OutgoingContractOutcome

Preimage

PreimageDecryptionShare

Share to decrypt an EncryptedPreimage

__FieldVisitor ğŸ”’

__FieldVisitor ğŸ”’

__FieldVisitor ğŸ”’

__FieldVisitor ğŸ”’

__FieldVisitor ğŸ”’

__FieldVisitor ğŸ”’

__Visitor ğŸ”’

__Visitor ğŸ”’

__Visitor ğŸ”’

__Visitor ğŸ”’

__Visitor ğŸ”’

__Visitor ğŸ”’

__Visitor ğŸ”’

__Visitor ğŸ”’

__Visitor ğŸ”’

Enums

Contract

A contract before execution as found in transaction outputs

ContractOutcome

Outcome of a contract. Only incoming contracts currently need to communicate anything back to the user (the decrypted preimage).

DecryptedPreimage

Possible outcomes of preimage decryption

FundedContract

A contract after execution as saved in the database

__Field ğŸ”’

__Field ğŸ”’

__Field ğŸ”’

__Field ğŸ”’

__Field ğŸ”’

__Field ğŸ”’

Constants

FIELDS ğŸ”’

FIELDS ğŸ”’

VARIANTS ğŸ”’

VARIANTS ğŸ”’

VARIANTS ğŸ”’

VARIANTS ğŸ”’

Traits

IdentifyableContract

Anything representing a contract which thus has an associated ContractId