Module fedimint_ln::contracts::incoming

Structs

FundedIncomingContract

The funded version of an IncomingContract contains the OutPoint of it’s creation. Since this kind of contract can only be funded once this out point is unambiguous. The out point is used to update the output outcome once decryption finishes.

IncomingContract

Specialized smart contract for incoming payments

IncomingContractOffer

OfferId

The hash of a LN incoming contract offer