Module fedimint_ln::contracts

Modules

account

incoming

outgoing

Structs

AccountContractOutcome

ContractId

The hash of a LN incoming contract

EncryptedPreimage

Threshold-encrypted Preimage

OutgoingContractOutcome

Preimage

PreimageDecryptionShare

Share to decrypt an EncryptedPreimage

Enums

Contract

A contract before execution as found in transaction outputs

ContractOutcome

Outcome of a contract. Only incoming contracts currently need to communicate anything back to the user (the decrypted preimage).

DecryptedPreimage

Possible outcomes of preimage decryption

FundedContract

A contract after execution as saved in the database

Traits

IdentifyableContract

Anything representing a contract which thus has an associated ContractId